Correct C++20 related warnings
Types of warnings:

warning: bitwise operation between different enumeration types is
deprecated [-Wdeprecated-enum-enum-conversion]

warning: compound assignment with ‘volatile’-qualified left operand is
deprecated [-Wvolatile]

warning: comparison of floating-point type with enumeration type is
deprecated [-Wdeprecated-enum-float-conversion]

warning: arithmetic between different enumeration types is deprecated
[-Wdeprecated-enum-enum-conversion]

+++ tests/pack/pack.cpp
@@ -147,8 +147,8 @@ struct TestPack {
       REQUIRE(min(p) == -1);
       REQUIRE(min(m, p.n, p) == -1);
       p[i] = p.n;
-      REQUIRE(max(p) == p.n);
-      REQUIRE(max(m, -1, p) == p.n);
+      REQUIRE(max(p) == (int)p.n);
+      REQUIRE(max(m, -1, p) == (int)p.n);
     }
   }
 
@@ -156,7 +156,7 @@ struct TestPack {
     Pack p_ref;
 
     // Make p_ref = [1 3 1 3 ... ]
-    vector_novec 
+    vector_novec
     for (int i = 0; i < PACKN; ++i) {
       p_ref[i] = 3-2*(i%2);
     }
@@ -199,7 +199,7 @@ struct TestPack {
     Pack p_ref;
 
     // Make p_ref = [1 3 1 3 ... ]
-    vector_novec 
+    vector_novec
     for (int i = 0; i < PACKN; ++i) {
       p_ref[i] = 3-2*(i%2);
     }
